Move some things and use primitive arrays.
