+++ src/main/kotlin/com/openlattice/data/storage/PostgresEntityDataQueryService.kt
@@ -3,7 +3,6 @@ package com.openlattice.data.storage
 import com.codahale.metrics.annotation.Timed
 import com.geekbeast.util.LinearBackoff
 import com.geekbeast.util.attempt
-import com.google.common.collect.Multimaps
 import com.openlattice.IdConstants
 import com.openlattice.analysis.SqlBindInfo
 import com.openlattice.analysis.requests.Filter
@@ -11,6 +10,7 @@ import com.openlattice.data.DeleteType
 import com.openlattice.data.WriteEvent
 import com.openlattice.data.storage.PostgresEntitySetSizesInitializationTask.Companion.ENTITY_SET_SIZES_VIEW
 import com.openlattice.data.storage.partitions.PartitionManager
+import com.openlattice.data.storage.partitions.getPartition
 import com.openlattice.data.util.PostgresDataHasher
 import com.openlattice.edm.type.PropertyType
 import com.openlattice.postgres.*
@@ -904,7 +904,11 @@ class PostgresEntityDataQueryService(
         val numUpdates = hds.connection.use { conn ->
             val ps = conn.prepareStatement(zeroVersionsForEntitiesInEntitySet)
 
-            entityKeyIds.groupBy { getPartition(it, partitions) }
+            entityKeyIds.groupBy {
+                getPartition(
+                        it, partitions
+                )
+            }
                     .forEach { (partition, rawEntityBatch) ->
                         val partitionsArr = PostgresArrays.createIntArray(conn, listOf(partition))
                         val idsArr = PostgresArrays.createUuidArray(conn, rawEntityBatch)
@@ -973,7 +977,11 @@ class PostgresEntityDataQueryService(
     ): WriteEvent {
         val entityKeyIdsArr = PostgresArrays.createUuidArray(conn, entityKeyIds)
 
-        val partitionsArr = PostgresArrays.createIntArray(conn, entityKeyIds.map { getPartition(it, partitions) })
+        val partitionsArr = PostgresArrays.createIntArray(conn, entityKeyIds.map {
+            getPartition(
+                    it, partitions
+            )
+        })
 
         val numUpdated = conn.prepareStatement(updateVersionsForEntitiesInEntitySet).use { ps ->
             ps.setLong(1, -version)
@@ -1013,7 +1021,11 @@ class PostgresEntityDataQueryService(
         return hds.connection.use { conn ->
             val propertyTypeIdsArr = PostgresArrays.createUuidArray(conn, propertyTypesToTombstone.map { it.id })
             val entityKeyIdsArr = PostgresArrays.createUuidArray(conn, entityKeyIds)
-            val partitionsArr = PostgresArrays.createIntArray(conn, entityKeyIds.map { getPartition(it, partitions) })
+            val partitionsArr = PostgresArrays.createIntArray(conn, entityKeyIds.map {
+                getPartition(
+                        it, partitions
+                )
+            })
 
             val numUpdated = conn.prepareStatement(updateVersionsForPropertyTypesInEntitiesInEntitySet()).use { ps ->
                 ps.setLong(1, -version)
